pull methods up
